Extract to functional groups
This is a little premature, but to avoid having another domain in the flat structure going to extract this early.

I do not recommend this style of organization for the rationale explained at https://www.appcontinuum.io.

For an evolving demo, this is a step to setup a move to a better grouping in the future.